+++ test/e2e/specs/admin/font-library.spec.js
@@ -42,8 +42,8 @@ test.describe( 'Font Library', () => {
 			const fileChooser = await fileChooserPromise;
 			// Provides coverage for https://github.com/WordPress/gutenberg/issues/59023.
 			await fileChooser.setFiles( [
-				'./test/e2e/assets/Exo2-Regular.woff',
-				'./test/e2e/assets/Exo2-SemiBoldItalic.woff2',
+				'./assets/Exo2-Regular.woff',
+				'./assets/Exo2-SemiBoldItalic.woff2',
 			] );
 
 			// Check fonts were installed.