+++ typings/style-imports/index.d.ts
@@ -7,3 +7,6 @@ declare module '*.module.scss' {
 	const classes: { [ key: string ]: string };
 	export default classes;
 }
+
+declare module '*.css';
+declare module '*.scss';